Few updates

+++ elyko/database/migrations/2016_03_21_231923_create_evaluations_table.php
@@ -11,7 +11,7 @@ class CreateEvaluationsTable extends Migration {
 			$table->integer('id')->unsigned();
 			$table->integer('uv_id')->unsigned();
 			$table->string('name');
-			$table->decimal('coefficient');
+			$table->float('coefficient');
 			$table->boolean('locked');
 			$table->primary('id');
 		});
